Pulling out safe_produce_new, introducing prepare.reduce_genbank re Programming:#555
safe_produce_new is a combination of derivative filename and mkstemp overwrite if necessary.

prepare file is for doing work with genbank and others trying to get
it ready for processing by the genbank processor. Right now it is
trying to produce a no-sequence version of the genbank file so it can
parse faster in the 2nd step of the wizard.