Starting implementng a scheme for passing configuration through and hashing it.
The goal is to pass configuration from the website to pynpact -- that
does the running. These options may alter the results, so when pynpact
builds filenames and inspects the filesystem for an existing version
of a given product there should be something in the filename to
distinguish a given set of parameters.

The idea is to supply to pynpact, from the website, a big dictionary
that can (optionally) contain all the data pynpact needs. In any given
function (e.g. extract) we pull out the keys relevant to that step and
build a sha1hash out of those keys and their associated values. That
hash becomes part of the filename for that product.

re Programming:#555 (0.75)