Refactoring the start page.
There is now one form on the page instead of several. I manipulated
the validation functions to match.

Buttons are now outside of the accordion. There is a new config
button, but as yet the actions aren't different.

The accordion activates whichever form was being displayed; autoheight
is off to take into account the entrez search which may return a
bunch.