Framework for processing page continuation
Pynpact: run_Allplots no longer calls all the other functions. There
is now a process function that calls the others in sequence and has an
optional soft_timeout. It won't start the next processing step if it
is already passed the soft_timeout, instead it will raise an
exception.

Spatweb: run url now renders a frame and some info for javascript to
start processing. We'll have jquery do calls based on that url,
getting back how far we've gotten. When it is done it will redirect to
the results frame.

The configuration in the mean time is stored in a filestore session.