Putting more emphasis on the processing page rather than results.
The idea is that since a lot of the information we want to display
already has a start on the processing page: use it instead of
redirecting away.

I have that all lined up, the next step is to make the softtimeout
report steps always, and include links to the files with that.

re #622 (1), #615 (0.5), #623 (0.5)