Working on config page help text.
The text is now a bit bigger, shows for select boxes, lasts until the
input element loses focus, and is using an updated version of the
qtip library.

I've also moved the add_help_text function inline since it was only
used here and the logic was a bit odd; it also now logs missing help
text.

re #626 (0.25)