Invert the sense of run_prediction to skip_prediction
Due to vagaries of HTML forms unchecked checkboxes aren't submitted so
it is somewhat tricky to detect when it is unchecked.

By reversing the config option we ensure that the default (to run the
prediction) doesn't require a key in the url/posts whereas if it is
switched away (skip_prediction) then that key will always be present.