First draft at configuring by b.p. coordinate rather than page number
On the config screen give the option of entering the starting
coordinate and ending coordinate instead of page numbers; these
default to 0 and length.

During the running of allplots we then use those instead of pages. The
number of pages is still calculated for the purpose of the status
report.

For the last page it doesn't cut off at the end_base, but rather
generates through the end of that page.