Taskqueue: setup loggin in tasks/processes to go to separate file
The logging for taskqueue itself is still just going to stderr but now
at the start of working on a task logging is configured to dump all
messages into a file that is named based on the task.

Later we can have the website read from that file in order to show
more literal task output status to the webbrowser.