Got email sending on task finished.
If an email address is supplied during config then we enqueue a
second job that waits on the first and sends an email when done. This
relies on the environement variable DJANGO_SETTINGS_MODULE being set
in the tqdaemon process... not particularly good.

If we do try and read settings in one of the pool processes there's a
good chance it will mess up the logging

Additionally I checked genome.ufl.edu; it doesn't look like the box is
setup to relay mail out so we still need to figure out an acceptable
mail server and set that in the settings.