Working on a django-tqdaemon.py
The idea is that it would set the settings module and initialize
logging appropriately and then do taskqueue.daemonize().

However, I don't particularly like that because we're bringing a *lot*
into the environment of the pool that isn't particularly needed. Most
importantly the logging configuration then causes the daemon logger to
capture pynpact information rather than letting it go to the task log
file.

So this commit is probably mostly worthless.