Removing usages of the softtimer; writing messages to stderr instead
In the current architecture we run everything in a background process
so the softtimer doesn't have much use. The messages are still good
for status information though so print them to new logger

The new logger 'statuslog' is for printing messages to be shown to the
client. It gets sent out through stderr which is collected into a file
by taskqueue.