ui: draw x axis
* split off graphing calculator into it's own file
* new `GC.xaxis` to calculate ticks + labels; it can't be self
  contained because the axis label width isn't know until it's drawn
  on a canvas
* more tests, but those seem to have declining usefulness. Minor
  tweaks to the graph layout break tests
* `drawAxisTicks` and `drawAxisLabels` on the grapher to share code
  between axes

There's some potential confusion between drawing code that assumes
we're offset (i.e. 0,0 is in a specific place) and that assumes we're
working with 0,0 being the top-left of the canvas.

The graphing calculator also has more redundant calculation than I'd
like. It's probably negligable, but making that a thicker object with
some state might simplify things.

refs Programming:#735(2)