Added a `Task` class that keeps track of the info for one task
This wraps the common triple in this code of (func, args, kwargs) we use
for tasks; as well as the tid and path and some methods to help out with
it.

Then add a bunch of unit tests to verify it's working and some
refactoring to make it easier to test and draw some boundaries better.

re #737 (2)