Sketch of frontend for asynchronous job pieces.
The old system had us starting the job asynchronously and then polling
for the results. We now want to be able to see some of the results files
as we're going so they are now treated all as separate jobs.

In this sketch we call `pynpact.main.process` which should return a
namedtuple of results files.

These are filenames/jobids (the two should be equivalent) that we will
then feed down to javascript to poll/fetch on individually.

re Programming:#737 (2)