graph: add `GraphingCalculator.zoom`
does the math and returns some measurements for zooming. Zooms while
retaining our position, so if you zoom in on gene 2000, after the zoom
the mouse cursor is still on 2000, but the rest of the graph has
shifted.

Has a huge comment describing the algrebra and number-line
approach. In a nutshell, we figure a percentage of where you clicked
pre-zoom, and use that to calculate the post-zoom measurements.

refs Programming:#735 (2.5)