Working on pynpact.main.process
In this iteration process takes a verb which is the target and gets
things done to make that happen.

The point of this is that computation is directed from
outside (e.g. javascript) and everything just knows what its
dependencies are and makes that happen.

The steps all got simpler, they no longer need to know anything about
the executor, they just yield the pieces they need done and the process
enqueues them with the executor.

This contains the start of a new parsing library to replace the old
prepare module.

re Programming:#737 (3.75)