Handle missing tasks
We were in a situation where an already completed task wasn't being
enqueued, but then the client was checking on the status of it and
getting task not found.

We solved this two fold:

The npactweb client which already knows that task IDs represent filenames
checks for the file presence directly

The pynpact enqueuer always enqueues the task even if it knows the job
is already done. The `producer` decorator has a check for the target
existing already so this shouldn't actually add too much load.

re Programming:#742