parse and process hits, put them in the graphspec
`GraphDealer` handles parsing hits, adding it to the graphspec that gets
passed to `Grapher` for actual drawing.

Reorganized to share code with the extracts parsing, have a few
functions that accepts a `key` param to tell what part of a dictionary
to upgrade. It's a little bit of meta-programming with magic strings,
but I don't think it's too bad.

refs Programming:#742 (.5)