Make python accept fasta files
This finishes the conversion from `pynpact.prepare` to the less eager
`pynpact.parsing`

This removes all the dependencies I could find in code where we required
a gbk file or any of its contents. We still try to treat it as a GBK but
fall back to treating it as a fasta file.

Most of the C programs are still scanning for the ORIGIN line so they
don't really work. The website, with the recent, updates isn't detecting
that error condition well; it just says "we're done, here's the
pdf" which is a dead link.

re #733