add `GraphConfig.toggleTrack` to turn them on/off
Not hooked to UI yet. Changes the internal representation for a track a
bit so we can keep an active bit, and that required a little bit of
refactoring, but added some more tests and feeling pretty good about
`GraphConfig` as a means to communicate the graph properties that don't
change per graph.

refs #846 (.25)