move more code off `GraphDealer`
* event handlers and `$watch`es go to `npactGraphPageCtrl`
* move more graph specifications go to `GraphConfig`
* delete dead code

The `$watch` takes out a lot of weird plumbing we had in `GraphDealer`
for managing mulitple redraw requests. Now we have a
`GraphConfig.refreshCommand` that compares old/new graph configs and
tells us what we need to do in order to refresh the view. This function
probably belongs elsewhere, but it's stuck with `GraphConfig` for now.

At the end of this, I think `GraphDealer` will disappear entirely.

refs #846 (2)