cache Y axis as a png
~48% improvement, at the expense of some rendering jitters. Keeps a
cache of `Image` objects using the json formatted headerspec as the
cache key. Right now, the headerSpec controls the y-axis/left area, but
is regenerated for every graph, so we can't use it and reference
equality as our hash key directly.

We're caching `Image` (as opposed to `Kinetic.Image`) because I'm not
sure what effects the `Kinetic.Stage.destroy` will have on the items in
the cache.

refs #864 (.5)