Removing ngInfiniteScroll
This plugin alerted when you neered the bottom but did not give us a
good approach to removing graphs that are no longer on the page. Since
finding if a graph is visible isn't very difficult get rid of
ngInfiniteScroll which is being in the way more than helping.

re #846