Watch graph box visibilty and only draw if so
We watch for scroll events in in the `npactGraphContainerCtrl` (on the
window) and `$apply` when we see that. I played around a bit with setting
up the event watch in the `npactGraph`, that seemed slower but my
testing procedure here was poor.

The `npactGraph` attempts to draw every `$digest` cycle but skips if it
has already drawn or isn't visible.

The function for determining visibility is on the container ctrl-- one
func copy on the hotpath instead of n, maybe better for tracing
compiler?

We don't try and track the number of visible graphs or how long it took
to finish a comple redraw cycle (since it is harder to tell how many that
will be).

I moved the resize and width tracking logic off of `GraphConfig` and
into `npactGraphContainer`, that isn't part of the Config panel.

re #846 (3)