Change visibility detection to be O(1)
Previously every graph line was checking if it was visible by running
`getBoundingClientRect` and doing some math based on that.

Profiling revealed that `getBoundingClientRect` was one of the top
CPU time consumers.

I changed it to not use `getBoundingClientRect` and instead use the
offset of the container, the current scrollY, the viewport height, and
the graph height to calculate the range of graphs that should be visible
at any given time, and then each graph can just check that.

re #846 (0.5)