Work on config management and driving from JS
Use the config data we get from the server to drive the profile size and
partitioning instead of trying to infer it from the downloaded nprofile
data. This allows us to start setting up the page earlier and is a much
logically cleaner setup.

This also allows us to drive the start and end base out of the control
panel on the left.

re #849 (0.5)
re #845 (0.75)