`GraphConfig.loadTrack` ensures hits are last
loading tracks manages the ordering of the track list so "hits" type
tracks are last (with test). This is to ensure they are as close to the
graph as possible.

The ordering is done at `loadTrack` time to minimize computation,
although we don't actually care about the ordering until graph rendering
time, so `Grapher` might have been a better place to do this.

refs Programming:#863 (.25)