Refactoring graph mouseEvents and option building
The options per graph used to include the scope so that we could emit
events on the scope. We then had to setup listeners on the scope to
listen for those events. We now just pass the onPan and onZoom functions
directly.

updateBaseOptions goes away and instead each watch that is detecting for
different changes updates the options with those changes directly.

re #847 (1)