Refactor Config tracks to allow replacement
Previously trying to load a track that was already defined was an error,
now it replaces the track.  This allows us to rerun server predictions
and replace the data.

I also ended up changing the "hits stay at bottom" code to use a weight
scheme so that we can better group tracks.