PredictionManager is now live
This is finally a fully working version of chaing the prediction
significance and getting new data.

The only quirk is that we didn't have a good way to invalidate the
tracks of the old prediction in the `Grapher` b/c it is keyed on name
and we are reusing the name: the solution, use a new name.

So now the tracks are 'Hits @0.001' and you can select both at once, but
by default it disables the old tracks when you select a new one.