draw background shading on profiles
* use `GraphingCalculator.shades`, draw those rectangles behind the
  profiles in `Grapher.profileGroup`
* new `npactConstants.profileShadeColor` option

There's still some futzy pixel rendering issues. If we tell the
background shades to go from y 0-100, then we start overwriting the
black border around the profile, which is drawn and cached "behind" the
profile group. Setting `Group.clipHeight` didn't seem to do anything
useful.

Tweaked some pixels to get it looking ok in most cases. Long term it
might be nice to do a major refactor of `Grapher` to get a better layout
of groups/layers/fastlayers and draw them in a better order.

ready #848 (.5)