Dynamically adjust step and window size
As the size of graph row shrinks we want to keep the resolution up so we
target about 200 data points per row.

We leave the window alone at 201 for now. This means the lines move very
slowly at high zoom, and are more jagged at low zoom.

I looked at just fixing the window to 4x the step (so 51 would go to
204), but I'm not sure that's any better.

re #845 (0.75)