Grapher redraw, simplify logic and reorder drawing
Previously we were testing for whether there were new options but in the
current incarnation there always are.

Now we render the leftLayer after the chartLayer giving the leftlayer an
implicitly higher z-index. This helps make shades that bleed over into
the leftLayer not obscure it.

I read around this code for a while, trying to get a better feel for it
and see if I could do the shades better, but this is what I got.

re #848 (1)