Better version of fixing margin at 0
Margin currently has a rendering error where the piece that should be in
the outside margin isn't clipped out appropriately.

We're less sure the margin has a lot of worth anyways though. Anytime
you click to drag a line you exceed it; and it otherwise is increasing
the rendering load by 20%.