Make window be dynamic
This makes the window size of the moving average calculation be dynamic
with respect to the step size we've chosen.

I came up with a log based graph (so it would scale sublinearly) that is
fixed around the traditional window of 201 for 10k range.

The step size continues to scale linearly (range/200).

Graphed:
http://www.wolframalpha.com/input/?i=plot+50*ln%28x%2F200%29+from+1000+to+100000

Samples:
Range   step   window
2000    9      111
10000   51     198
50000   249    276
At  2000 it is 111,
      // At 10000 it is 198
      // At 50000 it is 276

re #845 (0.25)