Fix onResize to not redraw everything
When we introduced the distinction between redraw and rebuild onresize
was still calling redraw. This would attempt to redraw everything that
had been drawn; if you had done a lot of scrolling and then resized that
would be a lot of drawing.

This changes redraw to a flag that indicates a redraw is required and
then draw always checks that flag as well as visibility.

In addition we don't draw directly from events but schedule it with a
timeout. This enables a basic amount of debouncing and makes the
interface in firefox notably better.

re #843 (2.25)