Restore promises in the grapher
In the previous refactoring the grapher.redraw was no longer returning a
promise. It now does.

Notes:
 * All draw calls are currently in redraw
 * only one of the three layer functions (genomeLayer)redraw calls returns a promise
 * in genomeLayer some things are promises, some aren't

This unevenness of some things return promises some things don't is a
bit hairy; I kind of think we need to move to, everything returns a
promise.

Also while this has coalesced and simplified the draw calls it has made
the partial rendering decrease somewhat. Now no drawing is done until
everything is there, which means you are waiting on some details a bit
more.

re #847 (0.5)