Readers: Use more promises to solve race condition
Previously loaded tracks only got added to the dictionary when they were
done parsing, but if someone tried to use it before then it got the
TrackNotFound error.

Instead we now stick the parsing promise into the dictionary and use it
to get the data when it comes time to slice.

re #847 (0.75)