headerSpecCalc => trackSizeCalc; rename headers as track
I found header to be confusing; does it mean the entry in the leftlayer
that is part of the key, or as it's been used the entire track row.

This also gets it into the graphing calculator so that the logic of
these heights is much closer to the calculation of the other heights.

re #917 (0.75)