Make height dynamic; rename styles everywhere
This is most of the way through big push to make height dynamic (now
mostly working) and try and make the graph building a little bit simpler
by using nested group coordinates.

I was originally hoping to use more of a vbox scheme for stacking things
vertically but since we want the left layer and the genomelayer to line
up vertically but be drawn separately it made more sense to go back to
calculating those coordinates ahead of time.

This patch also separates out style into something that the grapher and
calculator uses directly rather than getting lumped into one giant opts
dictionary that also has a lot of the data we're working with. I.e the
style is constant whereas the rest of the opts are more dynamic.

re #917 (2.5)