Simplify xAxis drawing
We used to:
 * figure out all the xaxis tick and label coordinates ahead of time
 * figure out all the shades as a separate thing except they are exactly
   tied to stops
 * Use some functions to convert from that representation to Kinetic.

This made it hard to trace through as the code came from lots of
different places.

Now it is perhaps slightly more redundant (the yaxis and xaxis code
doesn't go through the same code path), but I don't think that shared
code was actually a win for readability.