Further sketch of printing
This is a working iteration of causing all the rows to prerender and
capturing a promise for it. This works fairly well in chrome but in
firefox for me it slows down enough to trigger the script is running to
long warnings (even though there are liberal timeouts).

The next issue beyond it is that once all the rows are rendered and you
hit the browser print button the browser crashes under the load of
preparing to print.

Reading some resources like
http://www.html5rocks.com/en/tutorials/canvas/hidpi/ there's talk of the
backing store that the canvas is rendered on being higher dpi; not sure
if/how this is related to printing though. It seems like that may be
outdated according to
https://bugzilla.mozilla.org/show_bug.cgi?id=780362

So the next step to try is to, when printing, capture the image to
dataUri and put that as an image on the page replacing the canvas.

re #842 (1.5)