Don't render past the endBase
This includes GraphConfig into Grapher and uses it instead of trying to
pull all the options through the GraphContainer. Less translation, more
direct access.

The big impetus for this is that we need endBase and basesPerGraph
both in order to keep our render scale appropriate while not going
beyond the end base.

Update stops to not include a margin of its own, but instead call it
with the margin we've already calculated.

re #932 (1.5)