Setup RequestPDF button
This removes the FileManager and instead adds a DownloadsCtrl that is a
controller for that section of the page. It queries the prediction
manager and will respond to the button click for requesting a pdf to
make that happen.

The actual server side call to get the PDF isn't happening yet, but
that should be a relatively easy next step.

re #946 (0.5)