Working request pdf button
I moved the button down into the control panel since you can request
multiple times and that's where we have the config. Previously the
button went away once you requested it but we needed to add it back when
the config changed, that was easy to detect in the config panel, but now
we just always leave the button there so not sure this is actually a
win.

That gets traced through to the server with some improved number parsing
on the server and page title management.  This is so that we can pass
those variables from the client to the server reliably.

However there appears to be a bug in Allplots.c in how it reads the
NUCLEOTIDES. I verified that we are writing the correct string,
e.g. `C+G` as the second line of Allplots.def. However Dr. Brocchieri
appears to have editted the code in Allplots.c that reads that round
about [d7ff783] and now it is appearing in the generated PDF as `’#`.

Still TODO:
 * is the button in a good spot?
 * Figure out the NUCLEOTIDES title shenanigans
 * pass the x-tics interval from the client to the server, right now
  `1000` is hardcoded into `allplots.py:116`.
 * Adjust logic of `bp_per_page` to instead be based of `basesPerGraph`

re #946 (1.25)