Decrease track size, use paddingUnit
As we center stuff to track size it makes more sense to have that be
accurate of how big the track is. E.g. the label on the left is centered
vertically and if the track is taller than it actually renders than the
label is too low.

However that would make the tracks really jammed up against eachother so
in `trackSizeCalc` where we figure out the vertical layout use the
`paddingUnit` variable to put space back in between them.