Complete horizontal scrolling
This pays attention to the `keyup` in order to trigger a repaint after
we've stopped pressing.

This also pays attention to the `dragmove` to trigger the same offset
move left and right code so that when we drag with the mouse all the
bars on the page move at the same time.

I converted the box model for the graph to have the border outside the
width instead of inside as it made keeping track of scaleX easier

re #962 (2)