Draw ORF type tracks with varying styles
Previously we drew all the extracts, new orfs, as the same. We're now,
with the introduction of the 'modified' track varying the styles on this
track a bit more.

I went ahead and made new type names for the 'neworfs' and 'modified'
and use an expanded `trackStyles` dictionary in the constants to replace
the previous `trackHeights`. I didn't want to have to define all
combinations of styles and tracks so I used a defaults scheme.

re Programming:#970