Fixed tooltip hiding
Previously we tried to manually hide/destroy the tooltip when a number
of other events happened, this was being problematic in that we weren't
thorough enough.

I finally thought to look for qtip2 having a blur autohide and found
it. Now we have qtip2 take care of hiding and destroying the toolitp
whenever the tooltip loses focus.

In addition we no longer need to create the tooltip on the canvas
element (and this was problematic because it is the element focus that
matters). Instead we have a empty, non-visible div that the tooltips are
created on and tied to.

re #972