Change the tooltips positioning instruction
Originally we rendered it above if it was on the direct strand and below
if it was on the complement strand. This led to the tooltip sometimes
being off the top (without the ability to scroll up) and sometimes off
the bottom (can normally scroll down). Primarily though this was just
unintuitive as sometimes it was above and sometimes below and it was
hard to discern why.

We just removed this instruction and say always have the arrow tick be
at the `top center`, I.e. the tooltip is centered below where you click
the mouse which works reasonably well.

re #972