Fix row visibility calculation to match css
Previously you had to try and keep the constants in the JS file in sync
with the values in CSS. Even in the presence of giant comments saying
"please keep these in sync" they weren't.

Use jQuery's `outerHeight` to get the value exactly (at least when there
are actually graph rows and if there aren't, whatever).