Simplified Element-Width
Got rid of the magic constant that was used to align the graph with the
container. Now its simpler and the graph also expanded a little because
now there is more space. Got rid of padding in style.css for .graph
because it made the parent div smaller but now it is the same size as
the graph.