Cleanup and improve finding ORFs
Fixed a bug where a refined search wouldn't take you to a new location
because resultIndex hadn't changed.

I added a $timeout between the initial find results in the tracks and the
gather/filter/sort that processes the final resutls list. Since each of
these steps could go longish it's good to allow the browser a chance to
do stuff.

In the previous patch I was using the term 'hits' for results, that's
already used elsewhere in the system so renamed it here.

re #963 (0.5)