Resize the graph container to 7in when printing
The printing was looking fairly fuzzy because the print driver was
resizing the image down to fit on the page. This resize was done after
the text was rendered so it didn't look great.

This tries to resize the image down to a space that should fit on a page
well ahead of time so that the print driver doesn't have a chance to
mess it up.

re #973