Allow pip to fetch from PyPI
Originally I was trying to make sure that all python dependencies were
checked into the repo so that we could build even if PyPI went
away. That's probably still a good thing, but at least for the time
being we're going to back off on that stance and allow PyPi, which has
been very reliable, to do its job.