Pull out DDNA service, rework to fix offbyone error
We've been having a problem with the blast extracting where it is
consistently off by one.

I've poked at this a fair bit more and I think it is that the extract is
using the coordinate convention of GBK files that is 1 indexed and
ranges are inclusive at both ends. I think I've updated the code to
match this better and I pulled that bit of code out into a separate
service to make it much easier to test and verify this type of behavior.

re #736 (2)