Fix Caching Problems
Updated other pages to have files that will have modified time appended
onto the name so the browser will be forced to check for a new
version. Also added functionality for when not in debug mode so the
dynamic names will be based off of version number instead of time so it
will be slightly faster.