Fix scrolling bug, styling of Graph Controls
Added bootstrap classes to elements in the graph controls.
There was a bug when graph controls was longer than the content on the right, causing unexpected scroll behavior. Fixed this by adding a min-height onto the #progress div.
Finally, set a min-width on the entire mainContaier so that graph controls wouldn't shrink too much.