Add Mycoplasma functionality
Added checkbox for translating to Mycoplasma amino acids instead of
standard translation. Along with that the output of translater was
changed from x-www-form-urlencoded to json. Also, the data being passed
to the translator function in run.py was evaluated incorrectly because
of javascript so when the string "false" passed through it was evaluated
as true and all of the dna was being translated as Mycoplasma amino
acids. We fixed that by making sure the values were either true or
false.

re #981