Mycoplasma flag to acgt_gamma
Move the code for trying to detect if it is a mycoplasma from `acgt_gamma`
into `pynpact.parsing.mycoplasma`. acgt_gamma now accepts the flag from
the command line and the python makes sure it is provided.

The goal is to preserve the behavior of guessing, but be able to
override explicitly with the checkbox in the graph controls panel.

re #981 (0.75)