Disable the `trackSliceCache` in `Grapher`
This was using the name to indicate the track but with the mycoplasma
checkbox causing different track results for the same name the
trackslicecache is no longer valid.

Maybe we can store the url onto the track and use that as the key since
that already has a strong hash in it.

re #981 (0.25)