Remove unused libs from `requirements.txt`
`requirements.txt` is used during bootstrap to help make sure we get the
proper version of all libraries. With the switch from
npactweb+django+taskqueue to npactflask+flask+gevent the set of baseline
requirements have shifted.

I've not gone quite so far as to delete npactweb yet but take all of its
requirements out of the toplevel requirements installation.

re #915