Capture pynpact step output to log
In the gevent executor we want to capture this to the logfile whereas in
taskqueue we captured it to various .stderr files. It doesn't
particularly matter where it goes; the .stderr files were confusing,
having it mixed in the common log will also be confusing but at least
there's one place...

By putting a sigil value 'log' in the config and keying on that in
capproc to decide where stderr goes we make it somewhat configurable if
we ever want to switch back.

re #992 (2)