Setup for publishing to genome.ufl.edu
This is a bunch of changes to try and get us ready to actually publish
this new branch to genome.ufl.edu.

* Get rid of gunicorn; gevent can host the wsgi server fine. It doesn't
  handle the `X-Forwarded-For` header quite as nicely or run multiple
  worker processes but we don't need either of those.
* Config files (no residing in `etc/`)
  * apache.conf - reworked to use mod_proxy
  * init.d - LSB init.d script for starting supervisord
  * supervisord.conf config file for supervisor to actually launch our
    server

I've tried to setup the init.d script and supervisor so that Dr. Riva
could put it in place and then we can continue to administer the system
without requiring sudo access, but it will probably still be easier if
we have sudo.