Make tracks store data transparently
Previously the data was only available through the promise. This worked
pretty well because everything relied on that promise being fullfilled
anyways.

With the need to be able to edit tracks and save them back out (re #985)
we need to separate out the "wait for promise" and "access the data".