Graph->Grapher: pass $scope and $element directly
Previously we passed the non jQuery version of the element; but this
was just extra work unwrapping,wrapping, and again unwrapping it.

Previously $root was an injectable; this passes the $scope in more
directly. This also changes the `$broadcast('offset')` to
`$emit('offset')` which more accurately reflects what is going on and
sets us up for emitting more events as the graphrow will be usable under
multiple containers.

re #986