Improve static file cache policy
The `vSTATIC` versioning helps but only hits js resources we directly
require. Resources loaded by angular (template files) don't go through
that and get stuck in the browsers cache. This sets cache max-age to 0
when in DEBUG but re-enables it for the /raw/ directory under which
everything is strongly named.