Change acgt_gamma url to not return entire config dict
I found a race condition where if the `mycoplasma` flag (or any other
config) changes in the client while waiting for `acgt_gamma` to process
then it will overwrite the updated value when it comes back from the
server.  In the case of `mycoplasma` the server's value would update the
client triggering the `$watch` and caused a loop.

Now acgt_gamma just returns the couple keys that it has changed and the
JS deals with them directly.