Persist Zoomwindow onto querystring
this fleshes out the zoomwindow structure more and implements the share
link which persists it onto the querystring. This also makes
testing/development faster since i can just refresh the page and have
the window pop back up automatically.

I tried to standardize what the focus data is and created a new service
`FocusData` that is in charge of serializing and deserializing that to
the querystring.

re #986