big refactor of GraphContainer and ZoomWindow
I was trying to rework GraphContainer so it would handle some of the
lifting on the zoom popup window as well, but in the end that would have
made it pretty messy.

On the flip side I think the separation of concerns with on
npactGraphContainer now vs what is was is still a very positive move.