Work on zoomwindow object and rounding to 3
A bunch of work on the zoom window to try and make it use a consistent
object to represent the ORF we're editing that can then be saved.

I was working on rounding to a multiple of 3 but I now realize that is
incorrect: I need to round to a multiple of the phase.

re #985