Redo leftLayer cache as single cell
I found in profiling that calculating the cache key was really
expensive (the data we were working with changed since the code was
written).

I played a bit with changing the cache key to be simpler but ended up
deciding on a single-cell cache that gets cleared from the
container. This keeps more work out of the draw function.